# Speed up passing different `JaxSimModel` with same pytree structure to JIT-compiled functions