--- a/src/jaxsim/parsers/descriptions/joint.py
+++ b/src/jaxsim/parsers/descriptions/joint.py
@@ -41,7 +41,7 @@ class JointGenericAxis:
         return hash(self) == hash(other)
 
 
-@jax_dataclasses.pytree_dataclass
+@jax_dataclasses.pytree_dataclass(eq=False, unsafe_hash=False)
 class JointDescription(JaxsimDataclass):
     """
     In-memory description of a robot link.
@@ -97,23 +97,25 @@ class JointDescription(JaxsimDataclass):
 
     def __hash__(self) -> int:
 
+        from jaxsim.utils.wrappers import HashedNumpyArray
+
         return hash(
             (
                 hash(self.name),
-                hash(tuple(self.axis.tolist())),
-                hash(tuple(self.pose.flatten().tolist())),
+                HashedNumpyArray.hash_of_array(self.axis),
+                HashedNumpyArray.hash_of_array(self.pose),
                 hash(int(self.jtype)),
                 hash(self.child),
                 hash(self.parent),
                 hash(int(self.index)) if self.index is not None else 0,
-                hash(float(self.friction_static)),
-                hash(float(self.friction_viscous)),
-                hash(float(self.position_limit_damper)),
-                hash(float(self.position_limit_spring)),
-                hash((float(el) for el in self.position_limit)),
-                hash(tuple(np.atleast_1d(self.initial_position).tolist())),
-                hash(float(self.motor_inertia)),
-                hash(float(self.motor_viscous_friction)),
-                hash(float(self.motor_gear_ratio)),
+                HashedNumpyArray.hash_of_array(np.array(self.friction_static)),
+                HashedNumpyArray.hash_of_array(np.array(self.friction_viscous)),
+                HashedNumpyArray.hash_of_array(np.array(self.position_limit_damper)),
+                HashedNumpyArray.hash_of_array(np.array(self.position_limit_spring)),
+                HashedNumpyArray.hash_of_array(np.array(self.position_limit)),
+                HashedNumpyArray.hash_of_array(self.initial_position),
+                HashedNumpyArray.hash_of_array(np.array(self.motor_inertia)),
+                HashedNumpyArray.hash_of_array(np.array(self.motor_viscous_friction)),
+                HashedNumpyArray.hash_of_array(np.array(self.motor_gear_ratio)),
             ),
         )