--- a/src/jaxsim/parsers/descriptions/model.py
+++ b/src/jaxsim/parsers/descriptions/model.py
@@ -12,7 +12,7 @@ from .joint import JointDescription
 from .link import LinkDescription
 
 
-@dataclasses.dataclass(frozen=True)
+@dataclasses.dataclass(frozen=True, eq=False, unsafe_hash=False)
 class ModelDescription(KinematicGraph):
     """
     Intermediate representation representing the kinematic graph of a robot model.
@@ -28,7 +28,7 @@ class ModelDescription(KinematicGraph):
     fixed_base: bool = True
 
     collision_shapes: tuple[CollisionShape, ...] = dataclasses.field(
-        default_factory=list, repr=False, hash=False
+        default_factory=list, repr=False
     )
 
     @staticmethod