@diegoferigo 2024-06-14
Thanks @flferretti for the suggestion on more compact syntax. The original idea was to prevent evaluating all the conditions if one of the first ones is false. Your suggestion using `all` would work similarly only if its argument is a generator expression, which is not the case. I updated your suggestion to use chained `and`, Python stops early also in this case.